backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
PyMySQL==1.1.0
cryptography==41.0.5
python-dotenv==1.0.0
marshmallow==3.20.1
werkzeug==2.3.7
uuid==1.30
bcrypt==4.0.1